replace header.js
