--- src/components/header.module.css
@@ -1,17 +1,0 @@
-.nav {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  list-style-type: none;
-}
-
-.navlinks {
-  display: inline-block;
-  font-size: 1.5rem;
-}
-
-.logo {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
